backend/requirements.txt
fastapi
uvicorn
python-dotenv
cerebras-cloud-sdk
pdfminer.six
PyMuPDF
pydantic
python-multipart
python-docx
